Move `HotbarSlotType` to its own file
