+++ FFXIVClientStructs/FFXIV/Client/Game/UI/Hotbar.cs
@@ -5,7 +5,9 @@ namespace FFXIVClientStructs.FFXIV.Client.Game.UI;
 [StructLayout(LayoutKind.Explicit, Size = 0x08)]
 [VTableAddress("48 8D 05 ?? ?? ?? ?? 48 8B D9 48 89 01 48 81 C1 ?? ?? ?? ?? E8 ?? ?? ?? ?? 48 8D 05 ?? ?? ?? ?? C6 83", 3)]
 public unsafe partial struct Hotbar {
-    [FieldOffset(0x0), Obsolete("Use ClearCallback->vtbl")] public void* vtbl;
+    [FieldOffset(0x0), CExportIgnore, Obsolete("Use ClearCallback->vtbl")]
+    public void* vtbl;
+    
     [FieldOffset(0x0)] public RaptureHotbarModule.ClearCallback* ClearCallback;
 
     /// <inheritdoc cref="RaptureHotbarModule.ClearCallback.ResetAllHotbars"/>
